DataAPI+Desktop: Adapted DataModel and GeoData to fit together

+++ DataAPI/src/com/bdreiss/dataAPI/AilmentDataModel.java
@@ -84,7 +84,7 @@ public class AilmentDataModel extends AbstractDataModel{
 	}
 
 	@Override
-	public int count(String key, LocalDate date) throws EntryNotFoundException {
+	public float count(String key, LocalDate date) throws EntryNotFoundException {
 		return data.countAilment(key, date);
 	}
