+++ DataAPI/src/com/bdreiss/dataAPI/RemedyDataModel.java
@@ -78,7 +78,7 @@ public class RemedyDataModel extends AbstractDataModel{
 	}
 
 	@Override
-	public int count(String key, LocalDate date) throws EntryNotFoundException {
+	public float count(String key, LocalDate date) throws EntryNotFoundException {
 		return data.countRemedy(key, date);
 	}
