+++ DataAPI/src/com/bdreiss/dataAPI/SymptomDataModel.java
@@ -78,7 +78,7 @@ public class SymptomDataModel extends AbstractDataModel{
 	}
 
 	@Override
-	public int count(String key, LocalDate date) throws EntryNotFoundException {
+	public float count(String key, LocalDate date) throws EntryNotFoundException {
 		return data.countSymptom(key, date);
 	}
